Allow global configuration of default range of arrays and collections in prop tests
Fixes #4989
